Replace is_file and is_dir with file and dir
is_file and is_dir were redundant with filetests -f and -d, with
the additional benefit that -f is an actual plain file and not
just "not Path::Class::is_dir".  However, we implement -d as
is_dir since a Path::Class::Dir must be a directory anyway.  We
create both "dir" and "directory" aliases.