Add generated README.mkdn and META.json
This commit adds the README.mkdown and META.json files
generated during build.  The first is to provide a
nicer README for github.  The second makes it easier for
those without dzil to get dependencies using mymeta-requires.
README.PATCHING has been updated with an example of how
to do it.